% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_funcs.R
\name{plot.plotR}
\alias{plot.plotR}
\title{Evaluates plotR expressions}
\usage{
\method{plot}{plotR}(x, ...)
}
\arguments{
\item{x}{a plotR object created with \code{plotR()}}

\item{...}{optional additional arguments}
}
\description{
Evaluates plotR expressions
}
